Port bug review: fix the unsigned-index/64-bit family, unchecked GL locks, Bionic hazards; restore IXboxSystem
Sweep of the whole tree for siblings of the bugs found during bring-up,
each verified in context before fixing.

64-bit / unsigned arithmetic (family of the FillPreTessPatchIB crash):
  - VBRHeader.cpp (all three copies): SeekPointByTimeVBRI read m_pnToc[i-1]
    with unsigned i==0 when the accumulate loop never ran - multi-GB OOB
    read on 64-bit; also bound the loop by m_dwTableSize.
  - togles glmgr.cpp DrawRangeElements: (int) pointer truncation -> (intp),
    matching the desktop togl copy.
  - gl_lightmap.cpp: reinterpret_cast<unsigned int>(ptr) in AssertMsgs ->
    uintp (ill-formed on LP64; broke any assert-enabled build).
  - cmaterial.cpp: unsigned char m_VarCount wrapped at 256 vars, making
    realloc(p,0) free the array and then writing through it -> unsigned short.
  - cmatlightmaps.cpp: guard m_ImagePackers[Count()-1] on empty vector.

Unchecked buffer/texture locks (togles returns S_OK even when the mapping
failed, so every FAILED(hr) guard was dead on GLES):
  - texturedx8.cpp LockTexture/BlitSurfaceBits: null pBits -> fail/skip.
  - meshdx8.cpp CIndexBufferDx8/CVertexBufferDx8::Lock: null pLockedData
    takes the existing scratch-fallback path instead of handing out NULL.
  - meshdx8.cpp ModifyBeginEx: failed Modify left desc uninitialized (wild
    writes on displacement map load) - point desc at the dummy mesh.
  - meshdx8.cpp CopyStaticMeshIndexBufferToTempMeshIndexBuffer: guard read.
  - dynamicvb.h/dynamicib.h HandleLateCreation: guard the memcpy; keep the
    sysmem copy for retry, as the original comment suggested.
  - shaderapidx8.cpp frame-sync texture + CopyBitsFromHostSurface: guard.

Bionic / Android specifics:
  - launcher.cpp GrabSourceMutex: fall back to APP_DATA_PATH for the lock
    file - Android has no /tmp, so startup died with 'open(/tmp/...) failed'.
  - gameui/BasePanel.cpp: fclose(NULL) crash on audio-language change
    (fopen of /tmp always fails on Android; Bionic fclose derefs NULL).
  - strtools.cpp V_qsort_s: the LINUX arm silently returned WITHOUT SORTING;
    qsort_s.cpp is in the build - call it.
  - Stack-overflow class on ~1MB engine worker stacks: file-sized
    stackallocs to heap (physics_shared surfaceprops, util_shared ICE
    decode), LZSS compression work buffers (~100KB on LP64) to heap,
    enginetrace leaf lists (numleafs-sized) to CUtlVector, datamanager
    flush scratch to CUtlVector, studiodecal build buffers to the existing
    malloc path on Android.

Restore IXboxSystem (next startup blocker after the splash):
  - engine/wscript: stop excluding xboxsystem.cpp on Android - its PC-stub
    CXboxSystem provides XBOXSYSTEM_INTERFACE_VERSION, which server DLLInit,
    client CHLClient::Init and CGameUI::Initialize all hard-require.
  - steamdatagram_stubs_android.cpp: drop the now-duplicate null
    g_pXboxSystem definition.

+++ togles/linuxwin/glmgr.cpp
@@ -5039,11 +5039,11 @@ void GLMContext::DrawRangeElements(GLenum mode, GLuint start, GLuint end, GLsize
 	if ( pIndexBuf->m_bPseudo )
 	{
 		// you have to pass actual address, not offset
-		indicesActual = (void*)( (int)indicesActual + (int)pIndexBuf->m_pPseudoBuf );
-	} 
+		indicesActual = (void*)( (intp)indicesActual + (intp)pIndexBuf->m_pPseudoBuf );
+	}
 	if (pIndexBuf->m_bUsingPersistentBuffer)
 	{
-		indicesActual = (void*)( (int)indicesActual + (int)pIndexBuf->m_nPersistentBufferStartOffset );
+		indicesActual = (void*)( (intp)indicesActual + (intp)pIndexBuf->m_nPersistentBufferStartOffset );
 	}
 
 #if GLMDEBUG
